requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.10.0
accelerate>=0.20.0
peft>=0.4.0
numpy>=1.21.0
tqdm>=4.64.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
rouge-score>=0.1.2
codebleu>=0.1.0
black>=22.0.0
isort>=5.10.0